Change type of real numbers from float to double for numerical testing.

+++ rnn++/parser/basic_type.h
@@ -3,7 +3,7 @@ namespace rnn{
 namespace type{
 
 using idx_t = int64_t;
-using float_t = float;
+using float_t = double;
 using char_t = char;
 
 }//namespace rnn::type
